Python interface32

+++ retriever/lib/tools.py
@@ -5,42 +5,78 @@ scripts.
 
 """
 from __future__ import print_function
+
 from future import standard_library
+
 standard_library.install_aliases()
 
-from builtins import str
-from builtins import input
-from builtins import next
 import difflib
-import os
-import io
-from io import StringIO as newfile
-import warnings
-import unittest
-import shutil
-from decimal import Decimal
-from hashlib import md5
-
-from retriever import HOME_DIR, open_fr, open_fw, open_csvw, ENCODING, MODULE_LIST
-from retriever.lib.models import *
-import csv
 import json
-import xml.etree.ElementTree as ET
-warnings.filterwarnings("ignore")
+import platform
+import shutil
+import warnings
 
+from hashlib import md5
+from io import StringIO as newfile
+from retriever.lib.defaults import HOME_DIR, ENCODING
+from retriever.lib.scripts import MODULE_LIST
+from retriever.lib.models import *
+import xml.etree.ElementTree as ET
+
+warnings.filterwarnings("ignore")
 
 TEST_ENGINES = dict()
 
 
+def create_home_dir():
+    """Create Directory for retriever."""
+    current_platform = platform.system().lower()
+    if current_platform != 'windows':
+        import pwd
+
+    # create the necessary directory structure for storing scripts/raw_data
+    # in the ~/.retriever directory
+    for dir in (HOME_DIR, os.path.join(HOME_DIR, 'raw_data'), os.path.join(HOME_DIR, 'scripts')):
+        if not os.path.exists(dir):
+            try:
+                os.makedirs(dir)
+                if (current_platform != 'windows') and os.getenv("SUDO_USER"):
+                    # owner of .retriever should be user even when installing
+                    # w/sudo
+                    pw = pwd.getpwnam(os.getenv("SUDO_USER"))
+                    os.chown(dir, pw.pw_uid, pw.pw_gid)
+            except OSError:
+                print("The Retriever lacks permission to access the ~/.retriever/ directory.")
+                raise
+
+
 def name_matches(scripts, arg):
+    """Check for a match of the script in available scripts
+
+    if all, return the entire script list
+    if the exact script is available, return that script
+    if no exact script name detected, match the argument with keywords
+    title and name of all scripts and return the closest matches
+    """
+    arg = arg.strip().lower()
     matches = []
+
+    if not arg:
+        raise ValueError("No dataset name specified")
+
+    if arg == 'all':
+        return scripts
+
     for script in scripts:
-        if arg.lower() == script.name.lower(): return [script]
-        max_ratio = max([difflib.SequenceMatcher(None, arg.lower(), factor).ratio() for factor in (script.name.lower(), script.title.lower(), script.filename.lower())] +
-                        [difflib.SequenceMatcher(None, arg.lower(), factor).ratio() for factor in [keyword.strip().lower() for keywordset in script.keywords for keyword in keywordset]]
+        if arg == script.name.lower():
+            return [script]
+
+    for script in scripts:
+        max_ratio = max([difflib.SequenceMatcher(None, arg, factor).ratio() for factor in
+                         (script.name.lower(), script.title.lower(), script.filename.lower())] +
+                        [difflib.SequenceMatcher(None, arg, factor).ratio() for factor in
+                         [keyword.strip().lower() for keywordset in script.keywords for keyword in keywordset]]
                         )
-        if arg.lower() == 'all':
-            max_ratio = 1.0
         matches.append((script, max_ratio))
     matches = [m for m in sorted(matches, key=lambda m: m[1], reverse=True) if m[1] > 0.6]
     return [match[0] for match in matches]
@@ -54,44 +90,8 @@ def final_cleanup(engine):
 config_path = os.path.join(HOME_DIR, 'connections.config')
 
 
-def choose_engine(opts, choice=True):
-    """Prompts the user to select a database engine"""
-    from retriever.engines import engine_list
-
-    if "engine" in list(opts.keys()):
-        enginename = opts["engine"]
-    elif opts["command"] == "download":
-        enginename = "download"
-    else:
-        if not choice:
-            return None
-        print("Choose a database engine:")
-        for engine in engine_list:
-            if engine.abbreviation:
-                abbreviation = "(" + engine.abbreviation + ") "
-            else:
-                abbreviation = ""
-            print("    " + abbreviation + engine.name)
-        enginename = input(": ")
-    enginename = enginename.lower()
-
-    engine = Engine()
-    if not enginename:
-        engine = engine_list[0]
-    else:
-        for thisengine in engine_list:
-            if (enginename == thisengine.name.lower() or
-                    thisengine.abbreviation and
-                    enginename == thisengine.abbreviation):
-                engine = thisengine
-
-    engine.opts = opts
-    return engine
-
-
-def reset_retriever(scope):
-    """Remove stored information on scripts, data, and connections"""
-
+def reset_retriever(scope="all", ask_permission=True):
+    """Remove stored information on scripts, data, and connections."""
     warning_messages = {
         'all': "\nThis will remove existing scripts, cached data, and information on database connections. \nSpecifically it will remove the scripts and raw_data folders and the connections.config file in {}. \nDo you want to proceed? (y/N)\n",
         'scripts': "\nThis will remove existing scripts. \nSpecifically it will remove the scripts folder in {}.\nDo you want to proceed? (y/N)\n",
@@ -100,28 +100,35 @@ def reset_retriever(scope):
 
     path = os.path.normpath(HOME_DIR)
     warn_msg = warning_messages[scope].format(path)
-    confirm = input(warn_msg)
-    while not (confirm.lower() in ['y', 'n', '']):
-        print("Please enter either y or n.")
-        confirm = input()
+    if ask_permission:
+        confirm = input(warn_msg)
+        while not (confirm.lower() in ['y', 'n', '']):
+            print("Please enter either y or n.")
+            confirm = input()
+    else:
+        confirm = 'y'
     if confirm.lower() == 'y':
         if scope in ['data', 'all']:
-            shutil.rmtree(os.path.join(path, 'raw_data'))
+            if os.path.exists(os.path.join(path, 'raw_data')):
+                shutil.rmtree(os.path.join(path, 'raw_data'))
         if scope in ['scripts', 'all']:
-            shutil.rmtree(os.path.join(path, 'scripts'))
+            if os.path.exists(os.path.join(path, 'scripts')):
+                shutil.rmtree(os.path.join(path, 'scripts'))
 
 
 def json2csv(input_file, output_file=None, header_values=None):
-    """Convert Json file to CSV
-    function is used for only testing and can handle the file of the size
+    """Convert Json file to CSV.
+
+    Function is used for only testing and can handle the file of the size.
     """
-    file_out = open_fr(input_file, encode = False)
+    file_out = open_fr(input_file, encode=False)
     # set output file name and write header
     if output_file is None:
         output_file = os.path.splitext(os.path.basename(input_file))[0] + ".csv"
     csv_out = open_fw(output_file, encode=False)
     if os.name == 'nt':
-        outfile = csv.DictWriter(csv_out, dialect='excel', escapechar="\\", lineterminator='\n', fieldnames=header_values)
+        outfile = csv.DictWriter(csv_out, dialect='excel', escapechar="\\", lineterminator='\n',
+                                 fieldnames=header_values)
     else:
         outfile = csv.DictWriter(csv_out, dialect='excel', escapechar="\\", fieldnames=header_values)
     raw_data = json.loads(file_out.read())
@@ -135,8 +142,9 @@ def json2csv(input_file, output_file=None, header_values=None):
 
 
 def xml2csv(input_file, outputfile=None, header_values=None, row_tag="row"):
-    """Convert xml to csv
-    function is used for only testing and can handle the file of the size
+    """Convert xml to csv.
+
+    Function is used for only testing and can handle the file of the size.
     """
     file_output = open_fr(input_file, encode=False)
     # set output file name and write header
@@ -161,7 +169,7 @@ def xml2csv(input_file, outputfile=None, header_values=None, row_tag="row"):
 
 
 def getmd5(data, data_type='lines'):
-    """Get MD5 of a data source"""
+    """Get MD5 of a data source."""
     checksum = md5()
     if data_type == 'lines':
         for line in data:
@@ -196,8 +204,9 @@ def getmd5(data, data_type='lines'):
 
 
 def sort_file(file_path):
-    """Sort file by line and return the file
-    function is used for only testing and can handle the file of the size
+    """Sort file by line and return the file.
+
+    Function is used for only testing and can handle the file of the size.
     """
     file_path = os.path.normpath(file_path)
     input_file = open_fr(file_path)
@@ -212,8 +221,9 @@ def sort_file(file_path):
 
 
 def sort_csv(filename):
-    """Sort CSV rows minus the header and return the file
-    function is used for only testing and can handle the file of the size
+    """Sort CSV rows minus the header and return the file.
+
+    Function is used for only testing and can handle the file of the size.
     """
     filename = os.path.normpath(filename)
     input_file = open_fr(filename)
@@ -249,24 +259,28 @@ def sort_csv(filename):
 
 
 def create_file(data, output='output_file'):
-    """Writes a string to a file for use by tests"""
+    """Write lines to file from a list."""
     output_file = os.path.normpath(output)
     with open(output_file, 'w') as testfile:
-        testfile.write(data)
-        testfile.close()
+        for line in data:
+            testfile.write(line)
+            testfile.write("\n")
     return output_file
 
 
-def file_2string(input_file):
-    """return file contents as a string"""
+def file_2list(input_file):
+    """Read in a csv file and return lines a list."""
     input_file = os.path.normpath(input_file)
-    if sys.version_info >= (3, 0, 0):
-        input = io.open(input_file, 'rU')
-    else:
-        input = io.open(input_file, encoding=ENCODING)
 
-    obs_out = input.read()
-    return obs_out
+    if sys.version_info >= (3, 0, 0):
+        input_obj = io.open(input_file, 'rU')
+    else:
+        input_obj = io.open(input_file, encoding=ENCODING)
+
+    abs_list = []
+    for line in input_obj.readlines():
+        abs_list.append(line.strip())
+    return abs_list
 
 
 def get_module_version():
@@ -285,3 +299,15 @@ def get_module_version():
 
     scripts = sorted(scripts, key=str.lower)
     return scripts
+
+
+def set_proxy():
+    """Check for proxies and makes them available to urllib."""
+    proxies = ["https_proxy", "http_proxy", "ftp_proxy",
+               "HTTP_PROXY", "HTTPS_PROXY", "FTP_PROXY"]
+    for proxy in proxies:
+        if os.getenv(proxy):
+            if len(os.environ[proxy]) != 0:
+                for i in proxies:
+                    os.environ[i] = os.environ[proxy]
+                break
